[#1596] move `validate_eio_headers` back to `apps.py`
