+++ src/openzaak/components/documenten/import_utils.py
@@ -4,11 +4,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional
 
-from django.conf import settings
-from django.core.exceptions import ImproperlyConfigured
 from django.utils.functional import classproperty
-from djangorestframework_camel_case.util import underscore_to_camel
-from vng_api_common.serializers import GegevensGroepSerializer
 
 from openzaak.components.documenten.models import EnkelvoudigInformatieObject
 from openzaak.import_data.models import ImportRowResultChoices
@@ -289,56 +285,3 @@ class DocumentRow:
             "opmerking": self.comment,
             "resultaat": self.result,
         }
-
-
-def validate_eio_headers() -> None:
-    """
-    Validates that no serializer fields are missing in the `DocumentRow` class
-    """
-    if settings.CMIS_ENABLED:
-        return
-
-    from .api.serializers import EnkelvoudigInformatieObjectSerializer
-    from .import_utils import DocumentRow
-
-    excluded_fields = (
-        "locked",
-        "inhoud",
-        "bestandsdelen",
-        "begin_registratie",
-        "versie",
-        "url",
-    )
-
-    serializer = EnkelvoudigInformatieObjectSerializer()
-    fields = serializer.fields.items()
-    nested_fields = {
-        field_name: list(field.fields.keys())
-        for field_name, field in fields
-        if isinstance(field, GegevensGroepSerializer)
-    }
-    csv_nested_fields = set(
-        f"{field_parent}.{field_child}"
-        for field_parent, children in nested_fields.items()
-        for field_child in children
-    )
-
-    expected_fields = set(
-        underscore_to_camel(field)
-        for field in EnkelvoudigInformatieObjectSerializer.Meta.fields
-        if field not in (*excluded_fields, *nested_fields.keys())
-    )
-
-    expected_fields.update(csv_nested_fields)
-
-    document_row_fields = set(
-        field for field in DocumentRow.import_headers if field in expected_fields
-    )
-
-    missing_fields = expected_fields - document_row_fields
-
-    if missing_fields:
-        raise ImproperlyConfigured(
-            "The following fields are missing from the `DocumentRow` class: "
-            f"{','.join(missing_fields)}"
-        )
